Move wifi to it's own file
